Fix abstraction and make python backend work
