+++ modin/engines/python/pandas_on_python/frame/data.py
@@ -1,0 +1,7 @@
+from modin.engines.base.frame.data import BasePandasData
+from .partition_manager import PythonFrameManager
+
+
+class PandasOnPythonData(BasePandasData):
+
+    _frame_mgr_cls = PythonFrameManager
